book_generator/requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
supabase>=2.4.0
openai>=1.30.0
python-docx>=1.1.2
openpyxl>=3.1.2
gspread>=6.1.2
google-auth>=2.29.0
httpx>=0.27.0
pydantic>=2.9.0
pydantic-settings>=2.4.0
python-dotenv>=1.0.1
tenacity>=8.3.0
loguru>=0.7.2
celery>=5.4.0
redis>=5.0.4
